- Логика переключения экранов и старта игры перенесена в класс App
